Bug 1241454: Fix enum class storage of ExprType; r=bustage
UltraBlame original commit: 4bdcdd43b11d461b580e1cbe46b5eedd8513b733

+++ js/src/asmjs/WasmTypes.h
@@ -138,7 +138,7 @@ class Val
 
 
 
-enum class ExprType : uint8_t
+enum class ExprType : uint16_t
 {
     I32 = uint8_t(ValType::I32),
     I64 = uint8_t(ValType::I64),
